pyproject.toml
[tool.poetry]
name = "tecnic-bot"
version = "0.1.0"
description = ""
authors = ["OM GaminG,Vichu,Aashish,Anime.pdf /you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
disputils = "^0.1.3"
lavalink = "^3.1.2"
youtube_dl = "^2020.12.12"
ffmpeg = "^1.4"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
